Fix type errors and add type checking in CI (#20)
This PR contains a number of improvements to make the type checking pass and adds the check to the CI so we won't regress. The PR switches from IterableDataset to Dataset since IterableDatasets were never properly supported (since we take the length). We will fix this going forward and add proper support for IterableDatasets too.