Add database to store requests and background engine to process them (#28)
This is a first iteration of adding a database to store data about requests, and a background engine to process them. We are choosing sqlite as a DB right now to make things simple, but going forward for more serious use cases people might also be using postgres etc. Also we might want to introduce more types for the different requests and replies. Right now I'm keeping that open because it will likely evolve a decent amount.